Extract PackageModel diagnostics into separate file
